requirements.txt
streamlit
tensorflow
pillow
numpy
geopy
requests
beautifulsoup4
google-generativeai